requirements.txt
pyrogram
argparse
python-dotenv
python-slugify
iso-639
tgcrypto
ipinfo
pinata_python
tabulate
xmltodict
pywidevine
ciso8601
pytz
imdbpy
pinata
pyromod
pillow
beautifulsoup4
hachoir
jsbeautifier
